(hotfix devel)  Hotfixes for v11.0

See merge request alasca.cloud/tarook/tarook!2188

+++ actions/release-migrations/v11-01-yaml-hosts.sh
@@ -6,5 +6,10 @@ actions_dir="$(dirname "$0")/.."
 # shellcheck source=actions/lib.sh
 . "$actions_dir/lib.sh"
 
-notef "Removing obsolete state files..."
-run git rm -rf "$terraform_state_dir/rendered"
+if [ "${tf_usage:-true}" == 'true' ]; then
+    notef "Removing obsolete state files..."
+    run git rm -rf "$terraform_state_dir/rendered"
+
+    notef "Running Terraform stage to create output files"
+    run "$actions_dir/apply-terraform.sh"
+fi
